# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TTB Label Review PoC API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LabelUploadInput {
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  expectedBrandName?: string | null;
  /** @nullable */
  expectedClassType?: string | null;
  /** @nullable */
  expectedAlcoholContent?: string | null;
  /** @nullable */
  expectedNetContents?: string | null;
}

export interface LabelBatchInput {
  /** @nullable */
  sessionId?: string | null;
}

export type BeverageType = typeof BeverageType[keyof typeof BeverageType];


export const BeverageType = {
  SPIRITS: 'SPIRITS',
  WINE: 'WINE',
  MALT: 'MALT',
  UNKNOWN: 'UNKNOWN',
} as const;

export type OverallStatus = typeof OverallStatus[keyof typeof OverallStatus];


export const OverallStatus = {
  PASS: 'PASS',
  FAIL: 'FAIL',
  REVIEW: 'REVIEW',
} as const;

export type MatchStatus = typeof MatchStatus[keyof typeof MatchStatus];


export const MatchStatus = {
  PASS: 'PASS',
  FAIL: 'FAIL',
  NEEDS_REVIEW: 'NEEDS_REVIEW',
  NOT_APPLICABLE: 'NOT_APPLICABLE',
} as const;

export type FlagSeverity = typeof FlagSeverity[keyof typeof FlagSeverity];


export const FlagSeverity = {
  ERROR: 'ERROR',
  WARNING: 'WARNING',
  INFO: 'INFO',
} as const;

export interface FieldResult {
  /** @nullable */
  extractedValue?: string | null;
  /** @nullable */
  expectedValue?: string | null;
  matchStatus: MatchStatus;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  /** @nullable */
  failReason?: string | null;
  /** @nullable */
  isMandatory?: boolean | null;
}

export interface ComplianceFlag {
  field: string;
  severity: FlagSeverity;
  message: string;
}

export interface SameFieldOfVisionResult {
  compliant: boolean;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  /** @nullable */
  detectedOnPanel?: string | null;
  missingFromPanel?: string[];
  singleImageWarning: boolean;
}

export interface LabelAnalysisResult {
  labelId: string;
  sessionId: string;
  fileName: string;
  beverageType: BeverageType;
  overallStatus: OverallStatus;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidenceScore: number;
  brandName: FieldResult;
  classType: FieldResult;
  alcoholContent: FieldResult;
  netContents: FieldResult;
  governmentWarning: FieldResult;
  bottlerProducer: FieldResult;
  countryOfOrigin?: FieldResult | null;
  sameFieldOfVision?: SameFieldOfVisionResult | null;
  labelLanguage: FieldResult;
  prohibitedSurface: FieldResult;
  flags: ComplianceFlag[];
  processingMs: number;
  analyzedAt: string;
}

export interface BatchAnalysisResult {
  sessionId: string;
  totalCount: number;
  passCount: number;
  failCount: number;
  reviewCount: number;
  results: LabelAnalysisResult[];
}