# lib/api-zod/src/generated/types/batchAnalysisResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TTB Label Review PoC API
 * OpenAPI spec version: 0.1.0
 */
import type { LabelAnalysisResult } from './labelAnalysisResult';

export interface BatchAnalysisResult {
  sessionId: string;
  totalCount: number;
  passCount: number;
  failCount: number;
  reviewCount: number;
  results: LabelAnalysisResult[];
}