# lib/api-zod/src/generated/types/beverageType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TTB Label Review PoC API
 * OpenAPI spec version: 0.1.0
 */

export type BeverageType = typeof BeverageType[keyof typeof BeverageType];


export const BeverageType = {
  SPIRITS: 'SPIRITS',
  WINE: 'WINE',
  MALT: 'MALT',
  UNKNOWN: 'UNKNOWN',
} as const;