# lib/api-zod/src/generated/types/complianceFlag.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TTB Label Review PoC API
 * OpenAPI spec version: 0.1.0
 */
import type { FlagSeverity } from './flagSeverity';

export interface ComplianceFlag {
  field: string;
  severity: FlagSeverity;
  message: string;
}