# lib/api-zod/src/generated/types/fieldResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TTB Label Review PoC API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchStatus } from './matchStatus';

export interface FieldResult {
  /** @nullable */
  extractedValue?: string | null;
  /** @nullable */
  expectedValue?: string | null;
  matchStatus: MatchStatus;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  /** @nullable */
  failReason?: string | null;
  /** @nullable */
  isMandatory?: boolean | null;
}