# lib/api-zod/src/generated/types/flagSeverity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TTB Label Review PoC API
 * OpenAPI spec version: 0.1.0
 */

export type FlagSeverity = typeof FlagSeverity[keyof typeof FlagSeverity];


export const FlagSeverity = {
  ERROR: 'ERROR',
  WARNING: 'WARNING',
  INFO: 'INFO',
} as const;