# lib/api-zod/src/generated/types/labelAnalysisResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TTB Label Review PoC API
 * OpenAPI spec version: 0.1.0
 */
import type { BeverageType } from './beverageType';
import type { ComplianceFlag } from './complianceFlag';
import type { FieldResult } from './fieldResult';
import type { OverallStatus } from './overallStatus';
import type { SameFieldOfVisionResult } from './sameFieldOfVisionResult';

export interface LabelAnalysisResult {
  labelId: string;
  sessionId: string;
  fileName: string;
  beverageType: BeverageType;
  overallStatus: OverallStatus;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidenceScore: number;
  brandName: FieldResult;
  classType: FieldResult;
  alcoholContent: FieldResult;
  netContents: FieldResult;
  governmentWarning: FieldResult;
  bottlerProducer: FieldResult;
  countryOfOrigin?: FieldResult | null;
  sameFieldOfVision?: SameFieldOfVisionResult | null;
  labelLanguage: FieldResult;
  prohibitedSurface: FieldResult;
  flags: ComplianceFlag[];
  processingMs: number;
  analyzedAt: Date;
}