# lib/api-zod/src/generated/types/labelBatchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TTB Label Review PoC API
 * OpenAPI spec version: 0.1.0
 */

export interface LabelBatchInput {
  /** @nullable */
  sessionId?: string | null;
}