# lib/api-zod/src/generated/types/labelUploadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TTB Label Review PoC API
 * OpenAPI spec version: 0.1.0
 */

export interface LabelUploadInput {
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  expectedBrandName?: string | null;
  /** @nullable */
  expectedClassType?: string | null;
  /** @nullable */
  expectedAlcoholContent?: string | null;
  /** @nullable */
  expectedNetContents?: string | null;
}