# lib/api-zod/src/generated/types/overallStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TTB Label Review PoC API
 * OpenAPI spec version: 0.1.0
 */

export type OverallStatus = typeof OverallStatus[keyof typeof OverallStatus];


export const OverallStatus = {
  PASS: 'PASS',
  FAIL: 'FAIL',
  REVIEW: 'REVIEW',
} as const;