# lib/api-zod/src/generated/types/sameFieldOfVisionResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TTB Label Review PoC API
 * OpenAPI spec version: 0.1.0
 */

export interface SameFieldOfVisionResult {
  compliant: boolean;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  /** @nullable */
  detectedOnPanel?: string | null;
  missingFromPanel?: string[];
  singleImageWarning: boolean;
}